#11 Moved duplicate code for representing trigger and response to a new common parent class
